Move SpecHelpers to 'consensus' module
