+++ consensus/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':types')
   implementation project(':core')
+  implementation project(':crypto')
   implementation project(':pow')
 
   implementation 'com.google.guava:guava'
