+++ consensus/src/main/java/org/ethereum/beacon/consensus/SpecHelpers.java
@@ -1,5 +1,6 @@
-package org.ethereum.beacon.core;
+package org.ethereum.beacon.consensus;
 
+import org.ethereum.beacon.core.BeaconState;
 import org.ethereum.beacon.core.spec.ChainSpec;
 import org.ethereum.beacon.core.state.ShardCommittee;
 import org.ethereum.beacon.core.state.ValidatorRecord;
